+++ lib/solaar/ui/tray.py
@@ -55,7 +55,7 @@ def _create_menu(quit_handler):
 
     # per-device menu entries will be generated as-needed
 
-    no_receiver = Gtk.MenuItem.new_with_label(_('No Logitech device found'))
+    no_receiver = Gtk.MenuItem.new_with_label(_('No supported device found'))
     no_receiver.set_sensitive(False)
     menu.append(no_receiver)
     menu.append(Gtk.SeparatorMenuItem.new())
@@ -177,10 +177,8 @@ try:
 
     def _create(menu):
         _icons._init_icon_paths()
-        theme_paths = Gtk.IconTheme.get_default().get_search_path()
-
-        ind = AppIndicator3.Indicator.new_with_path(
-            'indicator-solaar', _icon_file(_icons.TRAY_INIT), AppIndicator3.IndicatorCategory.HARDWARE, theme_paths[0]
+        ind = AppIndicator3.Indicator.new(
+            'indicator-solaar', _icon_file(_icons.TRAY_INIT), AppIndicator3.IndicatorCategory.HARDWARE
         )
         ind.set_title(NAME)
         ind.set_status(AppIndicator3.IndicatorStatus.ACTIVE)